Refactory 3

+++ src/main/java/com/contaazul/coverage/cobertura/LineCoverager.java
@@ -1,5 +1,5 @@
 package com.contaazul.coverage.cobertura;
 
 public interface LineCoverager {
-	Integer getLineCoverage(int lineNumber);
+	Double getLineCoverage(int lineNumber);
 }
